fixup! Add boot steps configuration option
